Check for misisng arg in deploy_termcap.sh

+++ deploy_termcap.sh
@@ -2,5 +2,5 @@
 
 set -e
 
-scp dvtm.info $1:
-ssh $1 'mkdir -p .terminfo && tic dvtm.info && rm dvtm.info'
+scp dvtm.info ${1:?}:
+ssh ${1:?} 'mkdir -p .terminfo && tic dvtm.info && rm dvtm.info'
